+++ helpers/result.ts
@@ -64,3 +64,50 @@ export async function refresh_access_token(token: string, read: IRead, http: IHt
         console.log('Encountered error during refreshing access token:', refresh_response.data.error.message);
     }
 }
+
+export async function make_time_string( date: string, time: string, utc: number) : Promise<string>{
+
+    const date_string = date + 'T' + time + 'Z';
+    const new_date = new Date(date_string);
+    const datetime = new_date.toISOString();
+    const datetime_ms = datetime.split(".");
+
+    let final_time;
+    let decimal = Math.abs(utc - Math.floor(utc));
+    decimal = decimal * 60;
+    if (utc > 0) {
+                    utc = utc - (decimal / 60);
+                }
+    if (utc < 0) {
+                    utc = utc + (decimal / 60);
+                }
+
+    if (utc > 0 && utc < 10) {
+                    if (decimal == 0) {
+                        final_time = datetime_ms[0] + '+0' + utc + ':' + decimal + '0';
+                    } else {
+                    final_time = datetime_ms[0] + '+0' + utc + ':' + decimal;
+                    }
+                } else if (utc < 0 && utc > -10) {
+                    utc = utc * -1;
+                    if (decimal == 0) {
+                        final_time = datetime_ms[0] + '-0' + utc + ':' + decimal + '0';
+                    } else {
+                    final_time = datetime_ms[0] + '-0' + utc + ':' + decimal;
+                    }
+                } else if (utc >= 10) {
+                    if (decimal == 0) {
+                        final_time = datetime_ms[0] + '+' + utc + ':' + decimal + '0';
+                    } else {
+                    final_time = datetime_ms[0] + '+' + utc + ':' + decimal;
+                    }
+                } else {
+
+                    if (decimal == 0) {
+                        final_time = datetime_ms[0]  + utc + ':' + decimal + '0';
+                    } else {
+                    final_time = datetime_ms[0] + utc + ':' + decimal;
+                    }
+                }
+                return final_time;
+}
